LHCBPS-1421: added a first implementation of RunStampCheck
The optional service RunStampCheck can be instantiated (via
ApplicationManager::ExtSvc) to check if a special "run stamp"
condition exist for the current run.

The check relies on the run stamp condition being added only for
the closed IOVs for which the calibrations have been added to the
database, and no condition being present for the runs that cannot be
processed.