Made Decoders fail when ignoring lists for outputs entry in decoders
This may break a couple of well running tests, but they need to be fixed as they are time bombs.
Indeed, we still add lists in some places, silently ignored and tests were working simply because the default was ok.
Now when the default changes... I've spent hours in reverse engineering...