Don't cache the location when finding banks.
Such caching prevents decoder algorithms from using two separate raw
event locations.

Imagine an algorithm calling `findFirstRawBank`, and the last location
in the list of locations to search contains to requested bank, such that
the last location is cached. A subsequent call to `findFirstRawEvent`
will then return the cached location, even if the first location exists
(it just doesn't contain the raw bank requested during the call to
`firstFirstRawBank`).

A concrete example where this can happen is in PersitReco processing,
where the PersistReco object live in separate raw bank to the TCK, and
where both banks are required by the HLT decoder.