Fix calculation of number of banks in HltPackedDataWriter
The effect of the bug was that whenever the data size was an exact
multiple of MAX_PAYLOAD_SIZE (the portion that fits in a single
bank), an extra zero-length bank was added at the end. This should
have no impact on the decoding, appart from adding a few extra bytes.