Implemented proper and complete checking of counters in logs
Up to recently, counters were simply ignored in the log of the tests (except their header and thus the number of them).
I then implemented a way to check them, but you had to list the ones to check explicitely in the qmt files and the check was crude : pure string comparison.
This commit put in place systematic checking of all counters, with a default sensibility of 10-4. So differences are allowed as long as they are small enough compared to the value.
It will obviously break some tests in a first stage, some for a good reason, and they should be fixed, some others because the counters are naturally varying more than 10-4 (thing of memory usage for example). So This is to be followed by a second commit allowing to change the default sensibility on a per counter base.