Inability to open file RONLY when writing is not an error...
When creating a .cdb file from scratch, not being able to open it
for read is expected... Only when reading is not being able to
open the file RONLY an error.

Also, when creating from scratch, all lookups in the 'original' file
have to be skipped, but in way such that lookups in the write cache
are still done!