Modernize Primitives
Eliminate the need for Primitives.h to include itself after some
pre-processor manipulations. This is done by avoiding the dedicated
specialization for the 'void' functors (i.e. the case where there
is no argument). Instead, use variadic number of arguments, and
make both the case with and without argument a specialization.

In addition, implement `Min` and `Max` in terms of the generic
`BinaryOp` -- which implies moving the handling of 3 or more
functions into the `min` and `max` helpers in Math.h.