Simplify template argument specification
The argument type(s) of any transformed functor are, by construction,
equal to the argument type(s) of the input functor -- hence reflect
this in the way the types are specified, i.e. only require a single
Signature for UnaryOp, BinaryOp and BindBinary. Also, default the
return type of the transformed signature to the input signature.

Move the backwards compatible definions to the end of the file,
in order to avoid having to open and close the V2 namespace -- and,
in general, avoid mentioning V2 when possible.

Remove some decltype(auto) which are not necessary.