Commented crateODIN in Decoders' declarations
In Future branch, createODIN should not be added to the sequence as it usually fails when it tries to run.
The reason is that the ODIN bank is usually already present (whenever a condition is used anywhere to be precise) and the functional version of createODIN will try to overwrite it.
In the master branch, the old version was only creating the ODIN entry if not already there. But this behavior is not allowed anymore.
The conclusion is that by default createODIN should not appear in sequences in the new framework, except in very special cases (tiny tests) where no conditions are used.