Major change to IOV partitioning logic
To avoid injection of artificial IOV boundaries, the partitioning by
month now uses the "since" of the first entry in the month as "since"
for the month group (instead of midnight on the first day of the month).

The change comes with some deep refactoring of the code and with the
side effect that the "since" 0 is not anymore special in the grouping.