Fix DecisionBankKiller.AlwaysKeepBanks property
The property AlwaysKeepBanks uses an update handler, which was not
"used" in the constructor. This fixes it. See gaudi/Gaudi#16 for a
wider mitigation of this easy bug.

I've increased the coverage of the test, which was previously not
able to spot this bug.

While this code was in 2017-patches, it was never used in production.