Better reporting of failed counters
In a previous merge request I was adding display of counter differences even when the set of algorithms was different.
However, the counter differences were still not displayed if the set of counters for an algo was not the same.
This is now fixed and the display has been slightly improved also.