Remove timestamp from GOD output
This makes the stack build reproducible and therefore ccache hit efficiency
improves dramatically. This should save a lot of time in the nightlies
when there aren't many or deep changes being tested. It also saves time
locally for the same reason.

For example, for LHCb with a hot ccache we have
- Before this change: 570/1277 hits
- After this change: 1277/1277 hits