Modernize Relations
Apart from generic cleanups (eg. remove the 'inline' keyword when it
is not needed), the most important change is to change the way the
types used inside Relations are customized. When the traits class is
specialized for a given type, previously it had to specify _all_ types.
Instead here one _only_ specifies the types which need to be customized,
and then the remainder will be taken from the defaults (which may
involve previously customized types!). As a result, customizing the
types becomes much less work.