Exclude histogram headers in LHCb ref preprocessor
With LHCb!2260 counter and histogram summary tables are appended
explicitly to preprocessed references to create new refs.
Histogram header lines were not filtered by the preprocessor,
which leads to them appearing twice in the new refs, which in
turn makes the new refs invalid.
This MR fixes that by extending the list of exclusions.