Suffix line names with Decision in DecReports writer
Historically names in DecReports end with `Decision` and we should
keep this for consistency, see lhcb/Moore#130.

The logic of `ExecutionReportsWriter` is changed slightly such that
for each given line (composite node name) it looks for
`name + "Decision"` in the `ANNSvc`. This implicit map from node
name to decision name can be explicitly configured if we decide
it's worth it.